post code review 10/21

+++ Agar/Cell.cs
@@ -18,8 +18,8 @@ namespace Agar
     class Cell
     {
         
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
         public int Left { get; set; }
         public int Top { get; set; }
         //public Rect Bounds { get; set; }
